+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @keystone-alpha/test-utils
 
+## 2.2.4
+
+- Updated dependencies [33001656](https://github.com/keystonejs/keystone-5/commit/33001656):
+  - @keystone-alpha/adapter-knex@4.0.5
+  - @keystone-alpha/adapter-mongoose@4.0.3
+  - @keystone-alpha/keystone@12.0.0
+
 ## 2.2.3
 
 - Updated dependencies [e42fdb4a](https://github.com/keystonejs/keystone-5/commit/e42fdb4a):